refactor:  migrate to remove duplicate code
