+++ lib/functions.js
@@ -20,6 +20,16 @@ export const getYYYYMMDDHHMMSS = (date = new Date()) => {
   ].join('');
 };
 
+export const shuffleArray = (array) => {
+  for (let i = array.length - 1; i > 0; i--) {
+    const j = Math.floor(Math.random() * (i + 1));
+    const temp = array[i];
+    array[i] = array[j];
+    array[j] = temp;
+  }
+  return array;
+};
+
 export const readAsset = (assetPath) => {
   return fs.readFileSync(join(dirname, './assets', assetPath)).toString();
 };
@@ -41,7 +51,7 @@ export const pathConfig = async (options) => {
   
   let migrationsDir = join(PWD, 'migrations');
   let modelsDir = join(PWD, 'models');
-  let schemasDir = modelsDir;
+  let stateDir = modelsDir;
   let indexDir;
 
   if (sequelizerc['models-path']) modelsDir = sequelizerc['models-path'];
@@ -52,12 +62,12 @@ export const pathConfig = async (options) => {
   if (options.modelsPath) modelsDir = join(PWD, options.modelsPath);
   if (options.migrationsPath) migrationsDir = join(PWD, options.migrationsPath);
   if (options.indexFilePath) indexDir = join(PWD, options.indexFilePath);
-  if (options.statePath) schemasDir = join(PWD, options.statePath);
+  if (options.statePath) stateDir = join(PWD, options.statePath);
 
   return {
     modelsDir,
     migrationsDir,
-    schemasDir,
+    stateDir,
     indexDir,
     packageDir,
   };
