+++ lib/migration.js
@@ -1,161 +1,23 @@
 /* eslint-disable no-restricted-syntax */
 /* eslint-disable guard-for-in */
 /* eslint-disable no-continue */
-import Sequelize from 'sequelize';
-import hash from 'object-hash';
-import lodash from 'lodash';
-import deepDiff from 'deep-diff';
-
-import prettier from 'prettier';
-
 import { writeFileSync } from 'fs';
 import { join } from 'path';
+import lodash from 'lodash';
+
+import deepDiff from 'deep-diff';
+import prettier from 'prettier';
+
+import Sequelize from 'sequelize';
 
 import { readAsset, getYYYYMMDDHHMMSS } from './functions.js';
+import { getColumnTypeName, getDefaultValueType, parseIndex } from './parser.js';
 
-const { DataTypes: _DataTypes, ARRAY } = Sequelize;
 const { each, extend, clone, isEqual, template } = lodash;
-const { diff } = deepDiff;
 
 const { log } = console;
 
-const reverseSequelizeColTypeBase = (col, prefix = 'Sequelize.') => {
-  const attrName = typeof col.type.key === 'undefined' ? col.type : col.type.key;
-  const attrObj = col.type;
-  const options = col.type.options ? col.type.options : {};
-  const DataTypes = _DataTypes;
-
-  switch (attrName) {
-    // CHAR(length, binary)
-    case DataTypes.CHAR.key:
-      if (options.binary) return `CHAR.BINARY`;
-      return `CHAR(${options.length})`;
-
-    // STRING(length, binary).BINARY
-    case DataTypes.STRING.key:
-      return `STRING${options.length ? `(${options.length})` : ''}${
-        options.binary ? '.BINARY' : ''
-      }`;
-
-    // TEXT(length)
-    case DataTypes.TEXT.key:
-      if (!options.length) return `TEXT`;
-      if (typeof options.length === 'string') return `TEXT("${options.length.toLowerCase()}")`;
-      return `TEXT(${options.length.toLowerCase()})`;
-
-    // NUMBER(length, decimals).UNSIGNED.ZEROFILL
-    case DataTypes.NUMBER.key:
-    case DataTypes.TINYINT.key:
-    case DataTypes.SMALLINT.key:
-    case DataTypes.MEDIUMINT.key:
-    case DataTypes.BIGINT.key:
-    case DataTypes.FLOAT.key:
-    case DataTypes.REAL.key:
-    case DataTypes.DOUBLE.key:
-    case DataTypes.DECIMAL.key:
-    case DataTypes.INTEGER.key: {
-      let ret = attrName;
-      if (options.length) {
-        ret += `(${options.length}`;
-        if (options.decimals) ret += `, ${options.decimals}`;
-        ret += ')';
-      }
-
-      if (options.precision) {
-        ret += `(${options.precision}`;
-        if (options.scale) ret += `, ${options.scale}`;
-        ret += ')';
-      }
-
-      ret = [ret];
-
-      if (options.zerofill) ret.push('ZEROFILL');
-
-      if (options.unsigned) ret.push('UNSIGNED');
-
-      return `${ret.join('.')}`;
-    }
-
-    case DataTypes.ENUM.key:
-      return `ENUM('${options.values.join("', '")}')`;
-
-    case DataTypes.BLOB.key:
-      if (!options.length) return `BLOB`;
-      if (typeof options.length === 'string') return `BLOB("${options.length.toLowerCase()}")`;
-      return `BLOB(${options.length.toLowerCase()})`;
-
-    case DataTypes.GEOMETRY.key:
-      if (options.type) {
-        if (options.srid) return `GEOMETRY('${options.type}', ${options.srid})`;
-        return `GEOMETRY('${options.type}')`;
-      }
-      return 'GEOMETRY';
-
-    case DataTypes.GEOGRAPHY.key:
-      return 'GEOGRAPHY';
-
-    case DataTypes.ARRAY.key: {
-      const type = attrObj.toString();
-      const match = type.match(/(\w+)\[\]/);
-      const arrayType = match.length >= 1 ? prefix + match[1] : `${prefix}STRING`;
-      return `ARRAY(${arrayType})`;
-    }
-    case DataTypes.RANGE.key:
-      console.warn(`${attrName} type not supported, you should make it by`);
-      return attrObj.toSql();
-
-    // BOOLEAN, TIME, DATE, DATEONLY, HSTORE, JSONB
-    // UUID, UUIDV1, UUIDV4, VIRTUAL, INET, MACADDR
-    default:
-      return attrName;
-  }
-};
-
-const reverseSequelizeColType = (col, prefix = 'Sequelize.') => {
-  return `${prefix}${reverseSequelizeColTypeBase(col, prefix)}`;
-};
-
-const reverseSequelizeDefValueType = (defaultValue, prefix = 'Sequelize.') => {
-  if (typeof defaultValue === 'object') {
-    if (defaultValue.constructor && defaultValue.constructor.name) {
-      return { internal: true, value: prefix + defaultValue.constructor.name };
-    }
-  }
-
-  if (typeof defaultValue === 'function') return { notSupported: true, value: '' };
-
-  return { value: defaultValue };
-};
-
-const parseIndex = (idx) => {
-  delete idx.parser;
-  if (idx.type === '') delete idx.type;
-
-  const options = {};
-
-  if (idx.name) {
-    options.indexName = idx.name;
-    options.name = idx.name; // The name of the index. Default is __
-  }
-  // @todo: UNIQUE|FULLTEXT|SPATIAL
-  if (idx.unique) {
-    options.indicesType = 'UNIQUE';
-    options.type = 'UNIQUE';
-  }
-
-  if (idx.method) options.indexType = idx.type; // Set a type for the index, e.g. BTREE. See the documentation of the used dialect
-
-  if (idx.parser && idx.parser !== '') options.parser = idx.parser; // For FULLTEXT columns set your parser
-
-  idx.options = options;
-
-  idx.hash = hash(idx);
-
-  //    log ('PI:', JSON.stringify(idx, null, 4));
-  return idx;
-};
-
-const reverseModels = (sequelize, models) => {
+export const reverseModels = (sequelize, models) => {
   const tables = {};
 
   delete models.default;
@@ -175,7 +37,7 @@ const reverseModels = (sequelize, models) => {
         }
 
         if (property === 'defaultValue') {
-          const val = reverseSequelizeDefValueType(attributes[column][property]);
+          const val = getDefaultValueType(attributes[column][property]);
           if (val.notSupported) {
             log(`[Not supported] Skip defaultValue column of attribute ${model}:${column}`);
             delete attributes[column][property];
@@ -184,9 +46,7 @@ const reverseModels = (sequelize, models) => {
           attributes[column][property] = val;
         }
 
-        if (property === 'validate') {
-          delete attributes[column][property];
-        }
+        if (property === 'validate') delete attributes[column][property];
 
         // remove getters, setters...
         if (typeof attributes[column][property] === 'function') delete attributes[column][property];
@@ -207,12 +67,12 @@ const reverseModels = (sequelize, models) => {
             continue;
           }
           attributes[column].type = {
-            key: ARRAY.key,
+            key: Sequelize.ARRAY.key,
           };
         }
       }
 
-      let seqType = reverseSequelizeColType(attributes[column]);
+      let seqType = getColumnTypeName(attributes[column]);
 
       // NO virtual types in migration
       if (seqType === 'Sequelize.VIRTUAL') {
@@ -265,7 +125,7 @@ const reverseModels = (sequelize, models) => {
   return tables;
 };
 
-const parseDifference = (previousState, currentState) => {
+export const parseDifference = (previousState, currentState) => {
   const pushChangeColumn = (actions, tableName, df, options, depends) => {
     actions.push({
       actionType: 'changeColumn',
@@ -278,7 +138,7 @@ const parseDifference = (previousState, currentState) => {
 
   //    log(JSON.stringify(currentState, null, 4));
   const actions = [];
-  const difference = diff(previousState, currentState);
+  const difference = deepDiff.diff(previousState, currentState);
 
   for (const d in difference) {
     const df = difference[d];
@@ -492,7 +352,7 @@ const parseDifference = (previousState, currentState) => {
   return actions;
 };
 
-const sortActions = (actions) => {
+export const sortActions = (actions) => {
   const orderedActionTypes = [
     'removeIndex',
     'removeColumn',
@@ -544,13 +404,11 @@ const sortActions = (actions) => {
 
   // remove duplicate changeColumns
   for (let i = 0; i < actions.length; i++) {
-    if (isEqual(actions[i], actions[i - 1])) {
-      actions.splice(i, 1);
-    }
+    if (isEqual(actions[i], actions[i - 1])) actions.splice(i, 1);
   }
 };
 
-const getPartialMigration = (actions) => {
+export const getPartialMigration = (actions) => {
   const propertyToStr = (obj) => {
     const vals = [];
     for (const k in obj) {
@@ -585,13 +443,39 @@ const getPartialMigration = (actions) => {
     for (const attrName in attrs) {
       ret.push(`      "${attrName}": ${propertyToStr(attrs[attrName])}`);
     }
-    return ` { \n${ret.join(', \n')}\n     }`;
+    return `{ \n${ret.join(', \n')}\n     }`;
   };
 
   const addTransactionToOptions = (options) => {
     let ret = JSON.stringify({ ...options, transaction: '###TRANSACTION###' });
     ret = ret.replace('"transaction":"###TRANSACTION###"', 'transaction');
     return ret;
+  };
+  const writeAction = (action, pars = []) => {
+    pars = [
+      action.tableName,
+      ...pars,
+      action.options ? addTransactionToOptions(action.options) : '{ transaction }',
+    ];
+    let params = '';
+    for (const par of pars) {
+      if (typeof par === 'string' && par.substring(0, 1) !== '{' && par.substring(0, 1) !== '[')
+        params += `"${par}",`;
+      else params += `${par},`;
+    }
+    params = params.slice(0, -1);
+    return `{
+      fn: "${action.actionType}",
+      params: [
+        ${params}
+      ]
+    }`;
+  };
+
+  const writeConsoleOut = (action, pars = []) => {
+    const deps = action.depends ? `, deps: [${action.depends.join(', ')}]` : null;
+
+    return `${action.actionType}(${pars[0]})() => "${action.tableName}"${deps}`;
   };
 
   const commands = [];
@@ -602,126 +486,105 @@ const getPartialMigration = (actions) => {
     switch (action.actionType) {
       case 'createTable':
         {
-          const resUp = `{ fn: "createTable", params: [
-    "${action.tableName}",
-    ${getAttributes(action.attributes)},
-    ${addTransactionToOptions(action.options)}
-] }`;
+          const resUp = writeAction(action, [getAttributes(action.attributes)]);
           commands.push(resUp);
 
-          consoleOut.push(
-            `createTable "${action.tableName}", deps: [${action.depends.join(', ')}]`,
-          );
+          consoleOut.push(writeConsoleOut(action));
         }
         break;
 
       case 'dropTable':
         {
-          const res = `{ fn: "dropTable", params: ["${action.tableName}", { transaction }] }`;
-          commands.push(res);
+          const resUp = writeAction(action);
+          commands.push(resUp);
 
-          consoleOut.push(`dropTable "${action.tableName}"`);
+          consoleOut.push(writeConsoleOut(action));
         }
         break;
 
       case 'addColumn':
         {
-          const resUp = `{ fn: "addColumn", params: [
-    "${action.tableName}",
-    "${action.options && action.options.field ? action.options.field : action.attributeName}",
-    ${propertyToStr(action.options)},
-    { transaction }
-] }`;
-
+          const resUp = writeAction(action, [
+            action.options && action.options.field ? action.options.field : action.attributeName,
+            propertyToStr(action.options),
+          ]);
           commands.push(resUp);
 
-          consoleOut.push(`addColumn "${action.attributeName}" to table "${action.tableName}"`);
+          consoleOut.push(writeConsoleOut(action, [action.attributeName]));
         }
         break;
 
       case 'removeColumn':
         {
-          const res = `{ fn: "removeColumn", params: [
-    "${action.tableName}",
-    "${action.options && action.options.field ? action.options.field : action.columnName}",
-    { transaction }
-  ]
-}`;
-          commands.push(res);
+          const resUp = writeAction(action, [
+            action.options && action.options.field ? action.options.field : action.columnName,
+          ]);
+          commands.push(resUp);
 
           consoleOut.push(
-            `removeColumn "${
-              action.options && action.options.field ? action.options.field : action.columnName
-            }" from table "${action.tableName}"`,
+            writeConsoleOut(action, [
+              action.options && action.options.field ? action.options.field : action.columnName,
+            ]),
           );
         }
         break;
 
       case 'changeColumn':
         {
-          const res = `{ fn: "changeColumn", params: [
-    "${action.tableName}",
-    "${action.options && action.options.field ? action.options.field : action.attributeName}",
-    ${propertyToStr(action.options)},
-    { transaction }
-] }`;
-          commands.push(res);
+          const resUp = writeAction(action, [
+            action.options && action.options.field ? action.options.field : action.attributeName,
+            propertyToStr(action.options),
+          ]);
+          commands.push(resUp);
 
-          consoleOut.push(`changeColumn "${action.attributeName}" on table "${action.tableName}"`);
+          consoleOut.push(writeConsoleOut(action, [action.attributeName]));
         }
         break;
 
       case 'addIndex':
         {
-          const res = `{ fn: "addIndex", params: [
-    "${action.tableName}",
-    ${JSON.stringify(action.fields)},
-    ${addTransactionToOptions(action.options)}
-] }`;
-          commands.push(res);
-
           const nameOrAttrs =
             action.options && action.options.indexName && action.options.indexName !== ''
               ? `"${action.options.indexName}"`
               : JSON.stringify(action.fields);
-          consoleOut.push(`addIndex ${nameOrAttrs} to table "${action.tableName}"`);
+
+          const resUp = writeAction(action, [JSON.stringify(action.fields)]);
+          commands.push(resUp);
+
+          consoleOut.push(writeConsoleOut(action, [nameOrAttrs]));
         }
         break;
 
       case 'removeIndex':
         {
-          //                log(action)
           const nameOrAttrs =
             action.options && action.options.indexName && action.options.indexName !== ''
               ? `"${action.options.indexName}"`
               : JSON.stringify(action.fields);
 
-          const res = `{ fn: "removeIndex", params: [
-    "${action.tableName}",
-    ${nameOrAttrs},
-    { transaction }
-] }`;
-          commands.push(res);
+          const resUp = writeAction(action, [nameOrAttrs]);
+          commands.push(resUp);
 
-          consoleOut.push(`removeIndex ${nameOrAttrs} from table "${action.tableName}"`);
+          consoleOut.push(writeConsoleOut(action, [nameOrAttrs]));
         }
-
         break;
+
       default:
-      // code
+        // code
+        log('Error: action not specified');
     }
   }
 
   return { commands, consoleOut };
 };
 
-const getMigration = (upActions, downActions) => {
+export const getMigration = (upActions, downActions) => {
   const { commands: commandsUp, consoleOut } = getPartialMigration(upActions);
   const { commands: commandsDown } = getPartialMigration(downActions);
   return { commandsUp, commandsDown, consoleOut };
 };
 
-const writeMigration = (
+export const writeMigration = (
   revision,
   migration,
   migrationsDir,
@@ -757,12 +620,4 @@ const writeMigration = (
   writeFileSync(filename, content);
 
   return { filename, info };
-};
-
-export default {
-  writeMigration,
-  getMigration,
-  sortActions,
-  parseDifference,
-  reverseModels,
 };
